Backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
frozenlist==1.5.0
gotrue==2.11.3
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpie==3.2.3
httpx==0.28.1
hyperframe==6.1.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.1.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
postgrest==0.19.3
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.3.0
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.21.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
websockets==14.2
whitenoise==6.8.2
yarl==1.18.3
